+++ tests/phpunit/integration/plugin-loader.php
@@ -131,6 +131,7 @@ if ( ! function_exists( 'mai_get_version' ) ) {
 }
 
 require_once $plugin_root . '/lib/functions/utilities.php';
+require_once $plugin_root . '/lib/functions/options.php';
 require_once $plugin_root . '/lib/functions/widgets.php';
 require_once $plugin_root . '/lib/admin/upgrade.php';
 require_once $plugin_root . '/lib/admin/notices.php';